Vehicle Parking Management System in C++
The main objective of this C++ Project on Vehicle Parking Management System is 
to manage the details of parking fees, customers, parking slots, types, government 
tax, employees and manager salaries and vehicles entering and leaving the parking 
slots. This project will be totally built for administrative end and thus only the 
administrator is guaranteed the access.