SIMPLE CALCULATOR
This calculator is designed in javascript with the  help of html and css.
It is designed to calculate value through simple mathematical operations i.e DMAS.
It show us a  page which is like a calculator having different columns for calculations. And calculate values just by clicking on the given button.